+++ components/molecules/src/GlowLogo/index.ts
@@ -1,0 +1,1 @@
+export { GlowLogo } from "./GlowLogo";