+++ components/organisms/src/FooterLogo/index.ts
@@ -1,0 +1,1 @@
+export { FooterLogo } from "./FooterLogo";