+++ components/screens/src/LandingDaoboardScreen/index.ts
@@ -1,0 +1,1 @@
+export { LandingDaoboardScreen } from "./LandingDaoboardScreen";