+++ components/screens/src/LandingMoshScreen/index.ts
@@ -1,0 +1,1 @@
+export { LandingMoshScreen } from "./LandingMoshScreen";